requirements.txt
# Core frameworks
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.2
python-dotenv>=1.1.1
sqlalchemy==2.0.25
alembic==1.13.2
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Data & ML
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
spacy==3.7.4
PyMuPDF==1.24.9

# HTTP & async
requests>=2.32.3
aiofiles==24.1.0
aiohttp>=3.8.3,<4.0.0

# Pydantic
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.7.1

# Security & Sessions
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Latest CrewAI ecosystem (all compatible)
crewai>=0.80.0
crewai-tools>=0.12.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
